Split tdb into 3 files; should function the same
